Extract store to dedicated module
